BitmapApp/requirement.txt
Django==5.1.3
numpy==2.1.3
pillow==11.0.0